exercise2/requirements.txt
django
djangorestframework
django-filter
markdown
django-cors-headers
drf-yasg
